update: 対象がシンボリックリンクの場合に上書きでリンクを作成するよう変更

+++ dotfilesInstall.sh
@@ -8,45 +8,29 @@
 DOTFILES_DIRECTORY="dotfiles_vim"
 DOTFILES_LIST="vimrc gvimrc ideavimrc"
 for DOTFILE in $DOTFILES_LIST; do
-  if [ -e ~/.$DOTFILE ] || [ -L ~/.$DOTFILE ]; then
-    if [ ! -L ~/.$DOTFILE ]; then
-      # カスタマイズ済みのドットファイルを退避
-      mv ~/.$DOTFILE ~/.$DOTFILE.old
-    else
-      # シンボリックリンクを削除
-      rm ~/.$DOTFILE
-    fi
+  if [ -e ~/.$DOTFILE ] && [ ! -L ~/.$DOTFILE ]; then
+    # カスタマイズ済みのドットファイルを退避
+    mv ~/.$DOTFILE ~/.$DOTFILE.old
   fi
-  ln -s ~/$DOTFILES_DIRECTORY/_$DOTFILE ~/.$DOTFILE
+  ln -snf ~/$DOTFILES_DIRECTORY/_$DOTFILE ~/.$DOTFILE
 done
 
 # .directory
 # -eで存在確認（シンボリックリンクの場合はリンク先を評価する）
 # -Lはシンボリックリンクの確認
 # 両方チェックはシンボリックリンクのリンク切れへの対応のため
-if [ -e ~/.vim ] || [ -L ~/.vim ]; then
-  if [ ! -L ~/.vim ]; then
-    # カスタマイズ済みの.vimを退避する
-    mv ~/.vim ~/.vim.old
-  else
-    # ディレクトリのシンボリックリンクを削除する
-    rm ~/.vim
-  fi
+if [ -e ~/.vim ] && [ ! -L ~/.vim ]; then
+  # カスタマイズ済みの.vimを退避する
+  mv ~/.vim ~/.vim.old
 fi
+ln -snf ~/$DOTFILES_DIRECTORY/vimfiles ~/.vim
 
 mkdir -p ~/.config
-if [ -e ~/.config/nvim ] || [ -L ~/.config/nvim ]; then
-  if [ ! -L ~/.config/nvim ]; then
-    # すでにある設定を退避
-    mv ~/.config/nvim ~/.config/nvim.old
-  else
-    # シンボリックリンクを削除
-    rm ~/.config/nvim
-  fi
+if [ -e ~/.config/nvim ] && [ ! -L ~/.config/nvim ]; then
+  # すでにある設定を退避
+  mv ~/.config/nvim ~/.config/nvim.old
 fi
-
-ln -s ~/$DOTFILES_DIRECTORY/vimfiles ~/.vim
-ln -s ~/$DOTFILES_DIRECTORY/nvim ~/.config/nvim
+ln -snf ~/$DOTFILES_DIRECTORY/nvim ~/.config/nvim
 
 # install Dein.vim https://github.com/Shougo/dein.vim
 mkdir -p ~/.cache/dein
